(chore) Restructures file paths
