+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "mocha --reporter spec 'test/db/tests/*.js'",
+    "test": "mocha --reporter spec 'test/backend-tests/*.js'",
     "start": "gulp dev && nodemon server/server.js",
     "postinstall": "npm install -g bower && cd client && npm install && bower install"
   },
